Use polling model to ensure that event is in database.
There doesn't seem to be a reliable way to hook database writes by
the recorder. Since the recorder runs in its own thread, its database
session is unavailable in the main HA thread.

As a workaround, Variable polls the database whenever a tracked event
fires. Once the event is found in the database, then the var updates.

The manifest was also updated to ensure that recorder is loaded before
the var integration is loaded.